deployment/03_twitter_input/cloud-function_twitter_request_from_api/requirements.txt
# required pip packages
google-cloud-pubsub==2.9.0
tweepy
vaderSentiment